🎨 [UPDATE]: 게시글 수정-사진 삭제 수정

+++ src/main/java/KGUcapstone/OutDecision/domain/post/service/PostServiceImpl.java
@@ -213,7 +213,7 @@ public class PostServiceImpl implements PostService{
         for (String originImageUrl : originOptionImgList) {
             boolean existsInOptionsList = optionsList.stream()
                     .anyMatch(option -> originImageUrl.equals(option.getPhotoUrl()));
-            if (!existsInOptionsList) {
+            if (!existsInOptionsList && originImageUrl != null && !originImageUrl.isEmpty()) {
                 s3Service.deleteImage(originImageUrl);
             }
         }
